requirements.txt
fastapi
uvicorn
google-api-python-client
pydantic
requests